Separate resources into individual files
